Plugins for jenkins version 2.375.4 LTS
Plugins for jenkins version 2.375.4 LTS

There is known limitation in pipeline-utility-steps plugin which is
blocking us from running deployment pipelines. which is being fixed
by https://github.com/jenkinsci/pipeline-utility-steps-plugin/pull/206
until this pull request is merged, this PS is using specific
version of the plugin from this pull request.

quay.io plugin removed as this is not being used and has a vulnerability warning displayed

Change-Id: If9f14b6d65504eea25dc44456a54e2dfa1c4deb5